challenges/bgr/cookie-web/src/requirements.txt
Flask>=2.0
python-dotenv>=0.19